requirements.txt
numpy
scipy
pyyaml
pandas 
matplotlib 
seaborn 
vector 
awkward

--index-url https://download.pytorch.org/whl/cu121
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0

-f https://data.pyg.org/whl/torch-2.3.0+cu121.html
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

--index-url https://pypi.org/simple
torch-geometric
lightning
torchmetrics
PyYAML
kaggle
lightning[extra]
tensorboard
tensorboardX     